refactor: update data loading configuration and remove unused dataset file - Change the class path in train_ew_xfmr.yaml to reflect the new data structure, and delete the obsolete datasets.py file to streamline the codebase.
