+++ configs/training/train_ew_xfmr.yaml
@@ -92,7 +92,7 @@ model:
         freeze_snp_encoder: false
 
 data:
-  class_path: ml.data.dataloaders.TraceSeqEWDataloader
+  class_path: ml.data.eyewidth_data.TraceSeqEWDataloader
   init_args:
     data_dirs:
       pattern2_cowos_8mi: /proj/siaiadm/AI_training_data/D2D/UCIe_Trace/pattern2/cowos_8mi
