+++ ml/data/eyewidth_data.py
@@ -1,26 +1,248 @@
 import json
-import time
-import torch
+import psutil
 import pickle
+import random
 import numpy as np
-import pandas as pd
 from pathlib import Path
-from typing import Dict, List
-from sklearn.model_selection import train_test_split
+
+import torch
+import torch.nn.functional as F
+import torch.multiprocessing as mp
+from torch.utils.data import Dataset, DataLoader
+
 from lightning import LightningDataModule
 from lightning.pytorch.utilities import CombinedLoader
 from lightning.pytorch.utilities.rank_zero import rank_zero_info
 
 from ..utils.scaler import MinMaxScaler
-from .datasets import TraceDataset, TraceEWDataset, InferenceTraceDataset, InferenceTraceEWDataset, get_loader_from_dataset
 from .processors import CSVProcessor, TraceSequenceProcessor
 from simulation.parameters.bound_param import SampleResult, to_new_param_name
-from common.signal_utils import parse_snps, read_snp
+from common.signal_utils import parse_snps, read_snp, flip_snp
+
+def get_loader_from_dataset(
+    dataset: Dataset,
+    batch_size: int,
+    shuffle: bool = False
+):
+    drop_last = shuffle
+
+    # Optimize num_workers based on system capabilities
+    import os
+    cpu_count = os.cpu_count()
+    # Use fewer workers to avoid overwhelming the system
+    num_workers = min(4, cpu_count // 2) if cpu_count else 2
+
+    loader = DataLoader(
+        dataset=dataset,
+        batch_size=batch_size,
+        shuffle=shuffle,
+        pin_memory=True,
+        num_workers=num_workers,
+        persistent_workers=True if num_workers > 0 else False,
+        drop_last=drop_last,
+    )
+    return loader
+
+class SharedMemoryCache:
+    def __init__(self):
+        self.cache = mp.Manager().dict()
+
+    def get(self, key):
+        return self.cache.get(key, None)
+
+    def store(self, key, value):
+        if isinstance(value, tuple):
+            self.cache[key] = tuple(v.share_memory_() for v in value)
+        else:
+            self.cache[key] = value.share_memory_()
+
+class TraceEWDataset(Dataset):
+    def __init__(
+        self,
+        trace_seqs,
+        directions,
+        boundaries,
+        vert_snps,
+        eye_widths,
+        configs,
+        train=False,
+        ignore_snp=False
+    ):
+        super().__init__()
+
+        self.trace_seqs = torch.from_numpy(trace_seqs.copy()).float()
+        self.directions = torch.from_numpy(directions).int()
+        self.boundaries = torch.from_numpy(boundaries).float()
+        self.configs = configs
+
+        self.vert_snps = vert_snps
+        self.vert_cache = SharedMemoryCache()
+        self.ignore_snp = ignore_snp
+
+        self.eye_widths = torch.from_numpy(eye_widths).float()
+        self.repetition = self.boundaries.size(1)
+        self.train = train
+        
+        # Create dummy SNP data if ignoring SNPs
+        if self.ignore_snp:
+            # Create a small dummy SNP tensor: (2, F, P, P) where F=32, P=4 for minimal memory usage
+            self.dummy_snp = torch.zeros(2, 32, 4, 4, dtype=torch.complex64)
+
+    def __len__(self):
+        return len(self.trace_seqs) * self.repetition
+
+    def __getitem__(self, index):
+        seq_index = index // self.repetition
+        bnd_index = index % self.repetition
+
+        trace_seq = self.trace_seqs[seq_index]
+
+        if self.ignore_snp:
+            # Return dummy SNP data without loading files
+            vert_snp = self.dummy_snp.clone()
+        else:
+            # Retrieve the left and right snps
+            drv_vert_file, odt_vert_file = self.vert_snps[seq_index, bnd_index]
+            drv_vert_snp = self.vert_snp(drv_vert_file)
+            odt_vert_snp = self.vert_snp(odt_vert_file)
+            vert_snp = torch.stack((drv_vert_snp, flip_snp(odt_vert_snp)))
+
+        # Retrieve the boundary and eye width values
+        direction = self.directions[seq_index, bnd_index]
+        boundary = self.boundaries[seq_index, bnd_index]
+        eye_width = self.eye_widths[seq_index, bnd_index]
+        config = self.configs[seq_index, bnd_index]
+
+        if self.train and random.random() > 0.5 and not self.ignore_snp:
+            trace_seq, direction, eye_width, vert_snp = \
+                self.augment(trace_seq, direction, eye_width, vert_snp)
+        return trace_seq, direction, boundary, vert_snp, eye_width, config
+
+    def transform(self, seq_scaler, fix_scaler):
+        """Apply scaling transformations using semantic feature access."""
+        num = len(self.trace_seqs)
+
+        # Scale only the scalable features (geometry + additional features)
+        scalable_feats = TraceSequenceProcessor.get_scalable_features(self.trace_seqs)
+        feat_shape = scalable_feats.shape
+        scalable_feats_flat = scalable_feats.reshape(-1, feat_shape[-1])
+        scaled_feats = seq_scaler.transform(scalable_feats_flat).reshape(feat_shape)
+        
+        # Update only the scalable portion of the sequence
+        self.trace_seqs[:, :, TraceSequenceProcessor.get_scalable_slice()] = scaled_feats.float()
+
+        # Scale boundary features  
+        bound_dim = self.boundaries.size(-1)
+        scaled_boundary = fix_scaler.transform(self.boundaries).reshape(num, -1, bound_dim)
+        self.boundaries = scaled_boundary.float()
+        
+        return self
+
+    def load_snp(self, snp_file):
+        snp_data = self.vert_cache.get(snp_file)
+        if snp_data is None:
+            snp_data = read_snp(Path(snp_file)).s
+            snp_data = torch.from_numpy(snp_data).to(torch.complex64)
+            if self.check_memory(snp_data):
+                self.vert_cache.store(snp_file, snp_data)
+        return snp_data
+
+    def check_memory(self, data):
+        data_size = data.element_size() * data.nelement()
+        avail_mem = psutil.virtual_memory().available
+        return avail_mem > data_size
+
+    def augment(self, trace_seq, direction, eye_width, vert_snp, max_layer_idx=99, max_pos=10000.0):
+        seq_len = trace_seq.size(0)
+        seq_order = torch.arange(seq_len)
+        signal_indices = torch.where(trace_seq[:, 1] == 0)[0]
+        signal_order = -torch.ones_like(seq_order)
+        signal_order[signal_indices] = torch.arange(signal_indices.size(0))
+        pairs = torch.stack((seq_order, signal_order)).T
+
+        # Shuffle the sequence
+        pairs = pairs[torch.randperm(pairs.size(0))]
+        seq_order = pairs[:, 0]
+        signal_order = pairs[:, 1]
+        signal_order = signal_order[signal_order >= 0]
+        snp_order = torch.cat((signal_order, signal_order + len(signal_order)))
+
+        # Extract the shuffled sequence
+        trace_seq = trace_seq[seq_order]
+        eye_width = eye_width[signal_order]
+        direction = direction[signal_order]
+        vert_snp = vert_snp[:, :, :, snp_order]
+
+        # Augment layer values
+        max_layer = trace_seq[:, 0].max().int().item()
+        trace_seq[:, 0] += torch.randint(0, max_layer_idx - max_layer, (1,)).expand(seq_len)
+
+        # Augment x dim relative position values
+        max_spatial_dim = trace_seq[:, -2:].max(0).values
+        trace_seq[:, -2:] += (torch.rand((2,)) * (max_pos - max_spatial_dim)).unsqueeze(0).expand(seq_len, 2)
+
+        return trace_seq, direction, eye_width, vert_snp
+
+    def vert_snp(self, snp_file):
+        """Load and cache vertical SNP data."""
+        return self.load_snp(snp_file)
+
+class InferenceTraceEWDataset(Dataset):
+    def __init__(
+        self,
+        trace_seqs,
+        direction,
+        boundary,
+        drv_snp,
+        odt_snp
+    ):
+        super().__init__()
+
+        self.trace_seqs = torch.from_numpy(trace_seqs.copy()).float()
+        self.boundary = torch.from_numpy(boundary).float()
+        self.direction = torch.from_numpy(direction).int()
+
+        drv_snp = torch.from_numpy(drv_snp).to(torch.complex64)
+        odt_snp = torch.from_numpy(odt_snp).to(torch.complex64)
+        self.vert_snp = torch.stack((drv_snp, flip_snp(odt_snp)))
+        self.config = {
+            'boundary': boundary,
+            'directions': direction,
+            'snp_drv': drv_snp.name,
+            'snp_odt': odt_snp.name
+        }
+
+    def __len__(self):
+        return len(self.trace_seqs)
+
+    def __getitem__(self, index):
+        trace_seq = self.trace_seqs[index]
+        return trace_seq, self.direction, self.boundary, self.vert_snp, self.config
+
+    def transform(self, seq_scaler, fix_scaler):
+        """Apply scaling transformations using semantic feature access."""
+        num = len(self.trace_seqs)
+        
+        # Scale only the scalable features (geometry + additional features)
+        scalable_feats = TraceSequenceProcessor.get_scalable_features(self.trace_seqs)
+        feat_shape = scalable_feats.shape
+        scalable_feats_flat = scalable_feats.reshape(-1, feat_shape[-1])
+        scaled_feats = seq_scaler.transform(scalable_feats_flat).reshape(feat_shape)
+        
+        # Update only the scalable portion of the sequence
+        self.trace_seqs[:, :, TraceSequenceProcessor.get_scalable_slice()] = scaled_feats.float()
+
+        # Scale boundary features  
+        bound_dim = self.boundary.size(-1)
+        scaled_boundary = fix_scaler.transform(self.boundary).reshape(num, -1, bound_dim)
+        self.boundary = scaled_boundary.float()
+        
+        return self
 
 class TraceSeqEWDataloader(LightningDataModule):
     def __init__(
         self,
-        data_dirs: Dict[str, str],
+        data_dirs: dict[str, str],
         label_dir: str,
         batch_size: int,
         test_size: float = 0.2,
@@ -36,8 +258,8 @@ class TraceSeqEWDataloader(LightningDataModule):
         self.ignore_snp = ignore_snp
 
         # containers for datasets per "name"
-        self.train_dataset: Dict[str, TraceEWDataset] = {}
-        self.val_dataset: Dict[str, TraceEWDataset] = {}
+        self.train_dataset: dict[str, TraceEWDataset] = {}
+        self.val_dataset: dict[str, TraceEWDataset] = {}
 
     def setup(self, stage: str | None = None, padding_value: int = -1):
         # Scalers
@@ -210,7 +432,7 @@ class TraceSeqEWDataloader(LightningDataModule):
 class InferenceTraceSeqEWDataloader(LightningDataModule):
     def __init__(
         self,
-        data_dirs: List[str],
+        data_dirs: list[str],
         drv_snp: str,
         odt_snp: str,
         batch_size: int,
@@ -264,238 +486,4 @@ class InferenceTraceSeqEWDataloader(LightningDataModule):
         return CombinedLoader(
             get_loader_from_dataset(dataset=ds, batch_size=self.batch_size)
             for ds in self.predict_dataset
-        )
-
-class TraceSeqDataLoader(LightningDataModule):
-    def __init__(
-        self,
-        data_dir: Dict[str, str],
-        batch_size: int,
-        test_size: float = 0.2,
-        max_ports: int = 8,
-        scaler_path: str = None,
-    ):
-        super().__init__()
-        self.batch_size = batch_size
-        self.test_size = test_size
-        self.locate_input_csv_files(data_dir)
-        self.max_ports = max_ports
-        self.scaler_path = scaler_path
-        self.train_dataset = {}
-        self.val_dataset = {}
-
-    def locate_input_csv_files(self, data_dir):
-        self.input_csv_dir = {}
-        for name, csv_dir in data_dir.items():
-            rank_zero_info(f"Parsing data from {csv_dir}")
-            patterns = ["input for AI*.csv", "*AI input data*.csv"]
-            input_csv_dir = [
-                d for pattern in patterns for d in Path(csv_dir).glob(pattern)
-            ]
-            assert len(input_csv_dir) == 1, f"Cannot find input file in {csv_dir}"
-            self.input_csv_dir[name] = input_csv_dir[0]
-
-    def replace_zero_length_with_max(self, row):
-        max_val = row[row != 0].max()
-        row[row == 0] = max_val
-        return row
-
-    def parse_input_csv(self, csv_dir):
-        # Cast line types to integers
-        unique_types = ('S', 'G', 'D')
-        df = pd.read_csv(input_csv_dir, index_col=0, header=0)
-        df = df.replace({col: {v: i for i, v in enumerate(unique_types)} for col in df.columns[df.columns.str.contains('Type_')]})
-
-        def replace_zero_with_max(row):
-            max_val = row[row != 0].max()
-            row[row == 0] = max_val
-            return row
-
-        def load_and_preprocess_data(input_csv_dir):
-            df = pd.read_csv(input_csv_dir, index_col=0, header=0)
-            df = df.replace({col: {v: i for i, v in enumerate(unique_types)} for col in df.columns[df.columns.str.contains('Type_')]})
-            return df
-
-        def extract_features(df):
-            layer_idx = df.columns.str.contains('Layer_').nonzero()[0]
-            feat_dim = layer_idx[1] - layer_idx[0] - 3
-
-            width_idx = df.columns.str.contains('W_').nonzero()[0]
-            widths = df.iloc[:, width_idx]
-
-            height_idx = df.columns.str.contains('H_').nonzero()[0]
-            heights = df.iloc[:, height_idx]
-
-            length_idx = df.columns.str.contains('L_').nonzero()[0]
-            lengths = df.iloc[:, length_idx].apply(replace_zero_with_max, axis=1)
-
-            df.drop(df.columns[np.concatenate((width_idx, height_idx, length_idx))], axis=1, inplace=True)
-
-            return df, widths, heights, lengths, feat_dim
-
-        def calculate_dimensions(df, widths, heights, lengths):
-            layer_idx = df.columns.str.contains('Layer_').nonzero()[0]
-            layer_change = df.iloc[:, layer_idx].diff(axis=1).fillna(1).values.astype(bool)
-            layer_count = np.unique(df.iloc[:, layer_idx], return_counts=True)[1]
-            layer_count = layer_count / len(df).astype(int)
-
-            signal_layer_start = layer_count.cumsum()[np.where(layer_count != 1)[0] - 1][0]
-            signal_layer_end = layer_count.cumsum()[np.where(layer_count != 1)[0]][0]
-            signal_total_widths = widths.values.copy()
-            new_widths = signal_total_widths[..., None]
-            new_widths[:, signal_layer_end:] = signal_total_widths[..., None]
-            new_widths[:, signal_layer_end:] = signal_total_widths[..., None]
-            new_widths[:, signal_layer_end:] = signal_total_widths[..., None]
-            cum_widths = widths.cumsum(1, axis=1, fill_value=0).values
-            x_dim = cum_widths = np.repeat(cum_widths[layer_change].reshape((len(widths), -1)), layer_count, axis=1)
-
-            cum_heights = heights.values[layer_change].reshape((len(heights), -1)).cumsum(1)
-            cum_heights = np.roll(cum_heights, shift=1, axis=1)
-            cum_heights[:, 0] = 0
-            z_dim = np.repeat(cum_heights, layer_count, axis=1)
-
-            y_dim = np.dstack([np.zeros_like(lengths), lengths]).reshape((len(lengths), -1))
-
-            spatial_feats = np.dstack([new_widths, heights, lengths, x_dim, z_dim])
-
-            return spatial_feats, y_dim
-
-        df = load_and_preprocess_data(csv_dir)
-        df, widths, heights, lengths, feat_dim = extract_features(df)
-        spatial_feats, y_dim = calculate_dimensions(df, widths, heights, lengths)
-
-        new_input_arr = np.concatenate([df.values.reshape(len(df), -1, feat_dim), spatial_feats], axis=2)
-        new_input_arr = np.concatenate([np.repeat(new_input_arr, 2, axis=1), y_dim[..., None]], axis=2)
-
-        return new_input_arr, df.index
-
-    def setup(self, stage=None):
-        fit_scaler = True
-        try:
-            self.seq_scaler = torch.load(self.scaler_path)
-            rank_zero_info(f'Loaded scaler object from {self.scaler_path}')
-            fit_scaler = False
-        except (FileNotFoundError, AttributeError):
-            self.seq_scaler = MinMaxScaler()
-            rank_zero_info('Could not find scaler file in {self.scaler_path}, initiating new scaler object.')
-
-        for name, csv_dir in self.input_csv_dir.items():
-            input_arr, sample_idx = self.parse_input_csv(csv_dir)
-
-            # Read output snps
-            snp_dir = Path(csv_dir).parent
-            if Path(snp_dir, 'npz').exists():
-                snp_dir = snp_dir.joinpath("npz")
-            elif Path(snp_dir, 'snp').exists():
-                snp_dir = snp_dir.joinpath("snp")
-            snps, missing_idx = parse_snps(snp_dir, sample_idx.astype(np.float32))
-            for i in missing_idx:
-                input_arr = np.delete(input_arr, i, axis=0)
-
-            # Get number of ports
-            one_snp = read_snp(snps[0])
-
-            # Random choice N elements from input and snp
-            # seed = sum(ord(char) for char in str(csv_dir))
-            # np.random.seed(seed)
-            # indices = np.random.choice(input_arr.shape[0], 10, replace=False)
-            indices = np.arange(125)
-            # indices = np.arange(1000)
-            snps = [snps[i] for i in indices]
-            input_arr = input_arr[indices]
-            rank_zero_info(f'{name}| input_arr: {input_arr.shape}, SNP: ({len(snps)}*{one_snp.s.shape})')
-
-            # Split train and validation set
-            x_train, x_val, y_train, y_val = train_test_split(input_arr, snps, test_size=self.test_size)
-            x_train, x_val = train_test_split(input_arr, snps, test_size=self.test_size, shuffle=False)
-            # scale input but ignore layer and type features
-            if fit_scaler:
-                # For TraceSeqDataLoader, we scale all features except layer and type 
-                # (this dataset doesn't include spatial coordinates)
-                scalable_feats = input_arr[:, :, 2:]  # Skip layer and type
-                seq_feats_flat = scalable_feats.reshape(-1, scalable_feats.shape[-1])
-                self.seq_scaler.partial_fit(seq_feats_flat)
-
-            # store dataset
-            self.train_dataset[name] = TraceDataset(x_train, y_train, self.max_ports)
-            self.val_dataset[name] = TraceDataset(x_val, y_val, self.max_ports)
-
-        rank_zero_info(f'-----------Scaler-----------\n\
-samples seen: {self.seq_scaler.n_samples_seen_}\n\
-minimum: {self.seq_scaler.min_.round(decimals=2).tolist()}\n\
-maximum: {self.seq_scaler.max_.round(decimals=2).tolist()}')
-        for key in list(self.train_dataset.keys()):
-            self.train_dataset[name] = self.train_dataset[name].transform(self.seq_scaler, add_noise=True)
-            self.val_dataset[name] = self.val_dataset[name].transform(self.seq_scaler)
-
-        torch.save(self.seq_scaler, Path(self.trainer.logger.log_dir).joinpath("scaler.pth"))
-
-    def train_dataloader(self):
-        return CombinedLoader(
-            {
-                name: get_loader_from_dataset(
-                    dataset=dataset,
-                    batch_size=self.batch_size // len(self.train_dataset),
-                    shuffle=True,
-                )
-                for name, dataset in self.train_dataset.items()
-            },
-            'min_size'
-        )
-
-    def val_dataloader(self):
-        return CombinedLoader(
-            {
-                name: get_loader_from_dataset(
-                    dataset=dataset,
-                    batch_size=int(self.batch_size * 1.6 // len(self.val_dataset)),
-                )
-                for name, dataset in self.val_dataset.items()
-            },
-            'min_size'
-        )
-
-class InferenceTraceSeqDataLoader(TraceSeqDataLoader):
-    def __init__(
-        self,
-        data_dir: Dict[str, str],
-        batch_size: int,
-        max_ports: int,
-        scaler_path: str,
-    ):
-        super().__init__(data_dir, batch_size, max_ports=max_ports, scaler_path=scaler_path)
-        self.predict_dataset = {}
-
-    def setup(self, stage=None):
-        self.seq_scaler = torch.load(self.scaler_path)
-        rank_zero_info(f'Loaded scaler object from {self.scaler_path}')
-
-        for name, csv_dir in self.input_csv_dir.items():
-            input_arr, df = self.parse_input_csv(csv_dir)
-            # seed = sum(ord(char) for char in str(csv_dir))
-            # np.random.seed(seed)
-            # indices = np.random.choice(input_arr.shape[0], 100, replace=False)
-            input_arr = input_arr[:100]
-            rank_zero_info(f'{name}| input_arr: {input_arr.shape}')
-
-            # store dataset
-            self.predict_dataset[name] = InferenceTraceDataset(input_arr, self.max_ports)
-        rank_zero_info(f'-----------Scaler-----------\n\
-samples seen: {self.seq_scaler.n_samples_seen_}\n\
-minimum: {self.seq_scaler.min_.round(decimals=2).tolist()}\n\
-maximum: {self.seq_scaler.max_.round(decimals=2).tolist()}')
-        for key in list(self.predict_dataset.keys()):
-            self.predict_dataset[name] = self.predict_dataset[name].transform(self.seq_scaler)
-
-    def predict_dataloader(self):
-        return CombinedLoader(
-            {
-                name: get_loader_from_dataset(
-                    dataset=dataset,
-                    batch_size=self.batch_size * 2 // len(self.predict_dataset),
-                    shuffle=False,
-                )
-                for name, dataset in self.predict_dataset.items()
-            },
-            'sequential'
         )
